Move src/include/evict_inline.h -> src/evict/evict_inline.h
